Shift max auth db connections to auth advanced page
