+++ apps/studio/components/interfaces/Settings/Database/ConnectionPooling/ConnectionPooling.tsx
@@ -1,19 +1,20 @@
 import { zodResolver } from '@hookform/resolvers/zod'
 import { PermissionAction } from '@supabase/shared-types/out/constants'
-import { useParams } from 'common'
 import { capitalize } from 'lodash'
 import { Fragment, useEffect } from 'react'
 import { SubmitHandler, useForm } from 'react-hook-form'
 import { toast } from 'sonner'
 import z from 'zod'
 
+import { useParams } from 'common'
 import { useProjectContext } from 'components/layouts/ProjectLayout/ProjectContext'
 import AlertError from 'components/ui/AlertError'
 import { DocsButton } from 'components/ui/DocsButton'
+import { StringToPositiveNumber } from 'components/ui/Forms/Form.constants'
 import { FormActions } from 'components/ui/Forms/FormActions'
 import Panel from 'components/ui/Panel'
 import ShimmeringLoader from 'components/ui/ShimmeringLoader'
-import UpgradeToPro from 'components/ui/UpgradeToPro'
+import { useAuthConfigQuery } from 'data/auth/auth-config-query'
 import { useMaxConnectionsQuery } from 'data/database/max-connections-query'
 import { usePoolingConfigurationQuery } from 'data/database/pooling-configuration-query'
 import { usePoolingConfigurationUpdateMutation } from 'data/database/pooling-configuration-update-mutation'
@@ -39,47 +40,19 @@ import {
 import { Admonition } from 'ui-patterns'
 import { SESSION_MODE_DESCRIPTION, TRANSACTION_MODE_DESCRIPTION } from '../Database.constants'
 import { POOLING_OPTIMIZATIONS } from './ConnectionPooling.constants'
-import { useAuthConfigQuery } from 'data/auth/auth-config-query'
-import { useAuthConfigUpdateMutation } from 'data/auth/auth-config-update-mutation'
-import { useOrgSubscriptionQuery } from 'data/subscriptions/org-subscription-query'
-import { useSelectedOrganization } from 'hooks/misc/useSelectedOrganization'
-import { IS_PLATFORM } from 'lib/constants'
 
 const formId = 'connection-pooling-form'
-
-// This validator validates a string to be a positive integer or if it's an empty string, transforms it to a null
-const StringToPositiveNumber = z.union([
-  // parse the value if it's a number
-  z.number().positive().int(),
-  // parse the value if it's a non-empty string
-  z.string().min(1).pipe(z.coerce.number().positive().int()),
-  // transform a non-empty string into a null value
-  z
-    .string()
-    .max(0, 'The field accepts only a number')
-    .transform((v) => null),
-  z.null(),
-])
 
 const FormSchema = z.object({
   default_pool_size: StringToPositiveNumber,
   pool_mode: z.union([z.literal('transaction'), z.literal('session')]),
   max_client_conn: StringToPositiveNumber,
-  DB_MAX_POOL_SIZE: StringToPositiveNumber,
 })
 
 export const ConnectionPooling = () => {
   const { ref: projectRef } = useParams()
   const { project } = useProjectContext()
   const snap = useDatabaseSettingsStateSnapshot()
-  const organization = useSelectedOrganization()
-  const { data: subscription, isSuccess: isSuccessSubscription } = useOrgSubscriptionQuery({
-    orgSlug: organization?.slug,
-  })
-
-  const isTeamsEnterprisePlan =
-    isSuccessSubscription && subscription.plan.id !== 'free' && subscription.plan.id !== 'pro'
-  const promptTeamsEnterpriseUpgrade = IS_PLATFORM && !isTeamsEnterprisePlan
 
   const { data: addons } = useProjectAddonsQuery({ projectRef })
   const computeInstance = addons?.selected_addons.find((addon) => addon.type === 'compute_instance')
@@ -121,15 +94,12 @@ export const ConnectionPooling = () => {
     }
   )
 
-  const canUpdateConfig = useCheckPermissions(PermissionAction.UPDATE, 'custom_config_gotrue')
-
   const form = useForm<z.infer<typeof FormSchema>>({
     resolver: zodResolver(FormSchema),
     defaultValues: {
       pool_mode: poolingConfiguration?.pool_mode as 'transaction' | 'session',
       default_pool_size: poolingConfiguration?.default_pool_size as number | undefined,
       max_client_conn: poolingConfiguration?.max_client_conn as number | undefined,
-      DB_MAX_POOL_SIZE: authConfig?.DB_MAX_POOL_SIZE ?? null,
     },
   })
 
@@ -141,14 +111,11 @@ export const ConnectionPooling = () => {
             pool_mode: data.pool_mode,
             default_pool_size: data.default_pool_size,
             max_client_conn: poolingConfiguration?.max_client_conn,
-            DB_MAX_POOL_SIZE: authConfig?.DB_MAX_POOL_SIZE ?? null,
           })
         }
         toast.success('Successfully saved settings')
       },
     })
-
-  const { mutate: updateAuthConfig } = useAuthConfigUpdateMutation()
 
   const onSubmit: SubmitHandler<z.infer<typeof FormSchema>> = async (data) => {
     if (!projectRef) return console.error('Project ref is required')
@@ -159,20 +126,6 @@ export const ConnectionPooling = () => {
       default_pool_size: data.default_pool_size as number | undefined,
       pool_mode: data.pool_mode,
     })
-
-    if (data.DB_MAX_POOL_SIZE !== authConfig?.DB_MAX_POOL_SIZE) {
-      updateAuthConfig(
-        { projectRef, config: { DB_MAX_POOL_SIZE: data.DB_MAX_POOL_SIZE ?? undefined } },
-        {
-          onError: (error: Error) => {
-            toast.error(`Failed to update auth max direct connections: ${error?.message}`)
-          },
-          onSuccess: () => {
-            toast.success('Successfully updated auth max direct connections')
-          },
-        }
-      )
-    }
   }
 
   useEffect(() => {
@@ -181,7 +134,6 @@ export const ConnectionPooling = () => {
         pool_mode: poolingConfiguration?.pool_mode as 'transaction' | 'session',
         default_pool_size: poolingConfiguration?.default_pool_size as number | undefined,
         max_client_conn: poolingConfiguration?.max_client_conn as number | undefined,
-        DB_MAX_POOL_SIZE: authConfig?.DB_MAX_POOL_SIZE ?? null,
       })
     }
   }, [isSuccess, authConfig])
@@ -439,48 +391,8 @@ export const ConnectionPooling = () => {
                     </FormItem_Shadcn_>
                   )}
                 />
-                <FormField_Shadcn_
-                  control={form.control}
-                  name="DB_MAX_POOL_SIZE"
-                  render={({ field }) => (
-                    <FormItem_Shadcn_ className="grid gap-2 md:grid md:grid-cols-12 space-y-0">
-                      <FormLabel_Shadcn_ className="col-span-4 text-sm text-foreground-light">
-                        Max Direct Auth Connections
-                      </FormLabel_Shadcn_>
-                      <div className="col-span-8 space-y-2">
-                        {promptTeamsEnterpriseUpgrade && (
-                          <div className="mb-4">
-                            <UpgradeToPro
-                              primaryText="Upgrade to Team or Enterprise"
-                              secondaryText="Max Direct Auth Connections settings are only available on the Team Plan and up."
-                              buttonText="Upgrade to Team"
-                            />
-                          </div>
-                        )}
-                        <FormControl_Shadcn_ className="col-span-8">
-                          <Input_Shadcn_
-                            {...field}
-                            type="number"
-                            className="w-full"
-                            value={field.value || undefined}
-                            placeholder={field.value === null ? '10' : ''}
-                            disabled={!canUpdateConfig || isTeamsEnterprisePlan}
-                          />
-                        </FormControl_Shadcn_>
-                        <FormDescription_Shadcn_>
-                          Auth will take up no more than this number of connections from the total
-                          number of available connections to serve requests. These connections are
-                          not reserved, so when unused they are released. Defaults to 10
-                          connections.
-                        </FormDescription_Shadcn_>
-                        <FormMessage_Shadcn_ />
-                      </div>
-                    </FormItem_Shadcn_>
-                  )}
-                />
               </form>
             </Form_Shadcn_>
-            <div className="border-muted border-t"></div>
           </>
         )}
       </Panel>
