+++ apps/studio/pages/project/[ref]/auth/advanced.tsx
@@ -1,14 +1,9 @@
 import { PermissionAction } from '@supabase/shared-types/out/constants'
 
-import { AdvancedAuthSettingsForm, SmtpForm, ThirdPartyAuthForm } from 'components/interfaces/Auth'
+import { AdvancedAuthSettingsForm } from 'components/interfaces/Auth/AdvancedAuthSettingsForm'
 import AuthLayout from 'components/layouts/AuthLayout/AuthLayout'
 import DefaultLayout from 'components/layouts/DefaultLayout'
-import {
-  ScaffoldContainer,
-  ScaffoldDescription,
-  ScaffoldHeader,
-  ScaffoldTitle,
-} from 'components/layouts/Scaffold'
+import { ScaffoldContainer, ScaffoldHeader, ScaffoldTitle } from 'components/layouts/Scaffold'
 import NoPermission from 'components/ui/NoPermission'
 import { GenericSkeletonLoader } from 'components/ui/ShimmeringLoader'
 import { useCheckPermissions, usePermissionsLoaded } from 'hooks/misc/useCheckPermissions'
@@ -32,9 +27,7 @@ const PageLayout: NextPageWithLayout = () => {
         ) : !canReadAuthSettings ? (
           <NoPermission isFullPage resourceText="access your project's authentication settings" />
         ) : (
-          <>
-            <AdvancedAuthSettingsForm />
-          </>
+          <AdvancedAuthSettingsForm />
         )}
       </ScaffoldContainer>
     </>
